fixed bugs

+++ VolumeShot/Models/Exchange.cs
@@ -26,8 +26,8 @@ namespace VolumeShot.Models
                 OnPropertyChanged("Bet");
             }
         }
-        private int _requests { get; set; }
-        public int Requests
+        private double _requests { get; set; }
+        public double Requests
         {
             get { return _requests; }
             set
